backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
asyncpg==0.29.0

# Redis & Caching
redis==5.0.1
hiredis==2.2.3

# Task Queue
celery==5.3.4
kombu==5.3.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7

# MT4/MT5 Integration
MetaTrader5==5.0.48
websocket-client==1.6.4

# Telegram & Notifications
python-telegram-bot==20.3
requests==2.31.0

# Email & SMS
sendgrid==6.11.0
twilio==8.10.0

# WebSocket Real-time
websockets==12.0
python-socketio==5.10.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# ML & AI (Phase 3)
scikit-learn==1.3.2
tensorflow==2.15.0
joblib==1.3.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Monitoring & Logging
python-json-logger==2.0.7
prometheus-client==0.19.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30
